scripts: update runc version
since we run CI with develop branch of alibaba/runc, the runc version is
changing every day. but we need to use fix commit during the packaging.

therefore, for the v1.2.0 pouchcontainer, we use the specific commit in
the script. in the following pr, we can use env to modify the commit so
that we don't need to change it every time.

Signed-off-by: Wei Fu <fuweid89@gmail.com>